+++ packages/service-env/src/hooks.js
@@ -1,0 +1,4 @@
+const SERVICE_NAME = `env`;
+module.exports = {
+  SERVICE_NAME,
+};