+++ packages/service-env/src/index.js
@@ -1,3 +1,21 @@
-export { default } from './lib'
-export * from './lib/hooks'
-export * from './lib/get-env'
+/**
+ * Extends `process.env` with informations from local files:
+ *
+ * .env
+ * .env.local
+ * .env.[development|production|...]
+ * .env.[development|production|...].local
+ *
+ */
+
+const { SERVICE_NAME } = require('./hooks');
+const { initEnv } = require('./init-env');
+
+module.exports = ({ registerAction }) => {
+  registerAction({
+    hook: '$START',
+    name: SERVICE_NAME,
+    trace: __filename,
+    handler: initEnv,
+  });
+};