+++ packages/kitchensink/CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - Added examples for `@forrestjs/hooks`
 - Added examples for `@forrestjs/service-fastify`
+- Added `app01` examples for a complex app