+++ packages/service-env/CHANGELOG.md
@@ -1,2 +1,6 @@
 # Change Log
 
+## v1.21.0
+
+- Remove Babel
+- Update to `setContext` API