+++ packages/service-fastify/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Refactored to remove babel and the build step
 - Added support for a `/test` route that can be extended
 - Added globals utilities for E2E testing in Jest
+- Integrates with `service-env` to provide direct access to `getEnv` in the route handlers
 
 ## v1.13.1